Adding automatic browser opening and platform chack

+++ myVCF_GUI.py
@@ -1,6 +1,11 @@
 from Tkinter import *
 import os
+import sys
 import pip
+import webbrowser
+import time
+
+platform = sys.platform
 
 class App:
   def __init__(self, master):
@@ -20,13 +25,24 @@ class App:
     self.slogan.pack(side=LEFT)
   def run_app(self):
       print("Running myVCF...")
-      os.system("python manage.py runserver &")
-      print("Open browser and go to http://localhost:8000/")
+      if platform == "win32":
+          # Windows commands
+          pass
+      else:
+          os.system("python manage.py runserver &")
+      print("myVCF page is opening in the browser...")
+      time.sleep(2)
+      url = "http://localhost:8000/"
+      webbrowser.open(url,new=2)
       return True
 
   def stop_app(self):
       print("Stopping myVCF...")
-      os.system('pkill -f "python manage.py runserver"')
+      if platform == "win32":
+          # Windows command to
+          pass
+      else:
+          os.system('pkill -f "python manage.py runserver"')
       print "myVCF is shoutdown"
 
   def install_packages(self):
